- BLE + UI work

+++ s22_display/main/s22_model.h
@@ -8,14 +8,14 @@ extern "C" {
 #endif
 
 // Directly from BLE
-void set_voltage(uint8_t voltage);
-void set_speed(uint8_t speed);
-void set_current(int8_t amps);
-void set_temp(uint8_t temp);
+void set_voltage(uint16_t voltage);
+void set_speed(uint16_t speed);
+void set_current(int16_t amps);
+void set_temp(uint16_t temp);
 
 // Synthesized values?
-void set_pwm(uint8_t pwm);
-void set_batt_percent(uint8_t batt_percent);
+void set_pwm(uint16_t pwm);
+void set_batt_percent(uint16_t batt_percent);
 
 #ifdef __cplusplus
 } /*extern "C"*/
